Combine ws router into http router.
