+++ CMakeLists.txt
@@ -295,7 +295,6 @@ set(DROGON_SOURCES
     lib/src/Utilities.cc
     lib/src/WebSocketClientImpl.cc
     lib/src/WebSocketConnectionImpl.cc
-    lib/src/WebsocketControllersRouter.cc
     lib/src/YamlConfigAdapter.cc
     lib/src/drogon_test.cc)
 set(private_headers
@@ -327,7 +326,6 @@ set(private_headers
     lib/src/TaskTimeoutFlag.h
     lib/src/WebSocketClientImpl.h
     lib/src/WebSocketConnectionImpl.h
-    lib/src/WebsocketControllersRouter.h
     lib/src/FixedWindowRateLimiter.h
     lib/src/SlidingWindowRateLimiter.h
     lib/src/TokenBucketRateLimiter.h
