+++ lib/src/HttpControllerBinder.cc
@@ -1,6 +1,7 @@
 #include "HttpControllerBinder.h"
 #include "HttpResponseImpl.h"
 #include <drogon/HttpSimpleController.h>
+#include <drogon/WebSocketController.h>
 
 namespace drogon
 {
@@ -36,6 +37,43 @@ void HttpControllerBinder::handleRequest(
     auto &paramsVector = req->getRoutingParameters();
     std::deque<std::string> params(paramsVector.begin(), paramsVector.end());
     binderPtr_->handleHttpRequest(params, req, std::move(callback));
+}
+
+void WebsocketControllerBinder::handleRequest(
+    const HttpRequestImplPtr &req,
+    std::function<void(const HttpResponsePtr &)> &&callback) const
+{
+    std::string wsKey = req->getHeaderBy("sec-websocket-key");
+    wsKey.append("258EAFA5-E914-47DA-95CA-C5AB0DC85B11");
+    unsigned char accKey[20];
+    auto sha1 = trantor::utils::sha1(wsKey.c_str(), wsKey.length());
+    memcpy(accKey, &sha1, sizeof(sha1));
+    auto base64Key = utils::base64Encode(accKey, sizeof(accKey));
+    auto resp = HttpResponse::newHttpResponse();
+    resp->setStatusCode(k101SwitchingProtocols);
+    resp->addHeader("Upgrade", "websocket");
+    resp->addHeader("Connection", "Upgrade");
+    resp->addHeader("Sec-WebSocket-Accept", base64Key);
+    callback(resp);
+}
+
+void WebsocketControllerBinder::handleNewConnection(
+    const HttpRequestImplPtr &req,
+    const WebSocketConnectionImplPtr &wsConnPtr) const
+{
+    auto ctrlPtr = controller_;
+    assert(ctrlPtr);
+    wsConnPtr->setMessageCallback(
+        [ctrlPtr](std::string &&message,
+                  const WebSocketConnectionImplPtr &connPtr,
+                  const WebSocketMessageType &type) {
+            ctrlPtr->handleNewMessage(connPtr, std::move(message), type);
+        });
+    wsConnPtr->setCloseCallback(
+        [ctrlPtr](const WebSocketConnectionImplPtr &connPtr) {
+            ctrlPtr->handleConnectionClosed(connPtr);
+        });
+    ctrlPtr->handleNewConnection(req, wsConnPtr);
 }
 
 }  // namespace drogon
