+++ lib/src/HttpControllerBinder.h
@@ -2,6 +2,7 @@
 
 #include "ControllerBinderBase.h"
 #include "HttpRequestImpl.h"
+#include "WebSocketConnectionImpl.h"
 #include <drogon/HttpBinder.h>
 
 namespace drogon
@@ -27,6 +28,18 @@ class HttpControllerBinder : public ControllerBinderBase
     internal::HttpBinderBasePtr binderPtr_;
     std::vector<size_t> parameterPlaces_;
     std::vector<std::pair<std::string, size_t>> queryParametersPlaces_;
+};
+
+struct WebsocketControllerBinder : public ControllerBinderBase
+{
+    std::shared_ptr<WebSocketControllerBase> controller_;
+
+    void handleRequest(
+        const HttpRequestImplPtr &req,
+        std::function<void(const HttpResponsePtr &)> &&callback) const override;
+
+    void handleNewConnection(const HttpRequestImplPtr &req,
+                             const WebSocketConnectionImplPtr &wsConnPtr) const;
 };
 
 }  // namespace drogon
