+++ lib/src/impl_forwards.h
@@ -28,7 +28,6 @@ using WebSocketConnectionImplPtr = std::shared_ptr<WebSocketConnectionImpl>;
 class HttpRequestParser;
 class StaticFileRouter;
 class HttpControllersRouter;
-class WebsocketControllersRouter;
 class PluginsManager;
 class ListenerManager;
 class SharedLibManager;
